moving resource summary to resource manager
